feat(ui): polish header, tab bar, animations, and molecule presets
- Add HeaderDecorator: SO(2) rotating squares, floating point cloud dots, orbiting electron atom (pure CSS keyframes)
- Add UsageGuide: per-tab 3-step workflow strip with numbered badges
- Add Footer: references, tech stack, GitHub link
- Add MathTerm: CSS-only hover tooltips for SO(2), T-Net, E(3)
- Add ScrollProgress: fixed progress bar and scroll-to-top button
- Add tab-change fade+slide animation via .tab-enter keyframe
- Header: breadcrumb badge (Huong 01/03) and accent left-border per tab
- Tab buttons: icons (Circle/Box/FlaskConical) added
- Auto-load airplane point cloud on mount for Part 2
- EnergyResultCard: count-up animation with ease-out cubic
- ExampleMolecules: add Ethanol, Formaldehyde, Fluoromethane presets
- AtomTable/AtomRow: zebra striping, whitespace-nowrap headers, hidden spinners
- Translate all comments to professional English

+++ backend/services/pointnet_service.py
@@ -111,6 +111,13 @@ def load_models_if_needed() -> bool:
     else:
         print(f"[PointNet WARNING] Basic weights not found: {WEIGHTS_BASIC}")
 
+    # ── Ép về float32 để tránh dtype mismatch khi NequIP đổi global dtype ──
+    # NequIP's set_global_state() gọi torch.set_default_dtype(float64) globally.
+    # Nếu không ép kiểu tường minh, bias/weight của model sẽ là float64
+    # trong khi input tensor là float32 → crash khi inference.
+    _basic_model.float()
+    _full_model.float()
+
     # ── Set eval mode ──
     _basic_model.eval()
     _full_model.eval()
@@ -274,8 +281,9 @@ def classify(
     # ── Áp dụng perturbation TRƯỚC khi normalize/sample ──
     pts_raw = apply_perturbation(pts_raw, rotation_x, rotation_y, rotation_z, noise_level, drop_ratio)
 
-    pts_np = preprocess_point_cloud(pts_raw, num_points)   # (1, num_points, 3)
-    pts_tensor = torch.from_numpy(pts_np).float()          # tensor (1, num_points, 3)
+    pts_np = preprocess_point_cloud(pts_raw, num_points)       # (1, num_points, 3)
+    # Ép float32 tường minh — phòng trường hợp NequIP đã đổi default dtype thành float64
+    pts_tensor = torch.from_numpy(pts_np).to(torch.float32)    # tensor (1, num_points, 3)
 
     # ── Basic model inference ──
     basic_probs, basic_crit_idx = _run_inference(_basic_model, pts_tensor, is_full=False)
